root/backend/requirement.txt
asgiref==3.3.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-cors-headers==3.7.0
django-import-export==2.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.3.1
PyJWT==2.1.0
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.6
xlrd==2.0.1
xlwt==1.3.0
